+++ source/vjr/fonts.cpp
@@ -78,7 +78,7 @@
 		// See if there is an unused slot
 		//////
 			fontFirstUnused	= NULL;
-			lnLength		= strlen(tcFontName);
+			lnLength		= (s32)strlen(tcFontName);
 			for (lnI = 0; lnI < gFonts->populatedLength; lnI += sizeof(SFont))
 			{
 				// Grab the pointer
@@ -312,7 +312,7 @@
 
 
 		// Iterate from the current size upward for a maximum of 200 font point sizes, but also not more than 3x its current point size
-		lnTextLength	= strlen(tcText);
+		lnTextLength	= (s32)strlen(tcText);
 		lnFontBold		= ((tnFontBold) ? FW_BOLD : FW_NORMAL);
 		for (lnI = 0; lnI == 0 || (lnI < 200 && (s32)(tnFontSize + lnI) < (s32)(tnFontSize * 3) && lrc.bottom < (s32)((f32)tnHeightDesired * 1.25f) && lrc.right < (s32)((f32)tnWidthDesired * 1.25f)); lnI++)
 		{
